feat: updated design for landing page
