+++ packages/frontend/src/components/Home/AboutCharitySection/index.ts
@@ -1,0 +1,3 @@
+export { AboutPhoto } from './AboutPhoto';
+export { AboutCharitySection } from './AboutCharitySection';
+export { InfoCard } from './InfoCard';
