+++ packages/frontend/src/components/Home/HeroSection/index.ts
@@ -1,0 +1,1 @@
+export { HeroSection } from './HeroSection';
