+++ packages/frontend/src/components/Home/ShowInfoCard/index.ts
@@ -1,0 +1,1 @@
+export { ShowInfoCard } from './ShowInfoCard';
