+++ packages/frontend/src/components/Home/SponsorsSection/index.ts
@@ -1,0 +1,1 @@
+export { SponsorsSection } from './SponsorsSection';
