+++ packages/frontend/src/components/Home/index.ts
@@ -1,0 +1,4 @@
+export { AboutCharitySection } from './AboutCharitySection';
+export { HeroSection } from './HeroSection';
+export { ShowInfoCard } from './ShowInfoCard';
+export { SponsorsSection } from './SponsorsSection';
